requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
faiss-cpu
pypdf
python-dotenv